+++ Source/PulsarEngine/Settings/Settings.cpp
@@ -108,18 +108,18 @@ bool Mgr::HasTrophy(PulsarId id, TTMode mode) const {
     return this->HasTrophy(CupsConfig::sInstance->GetCRC32(id), mode);
 }
 
-u8 Mgr::GetSettingValue(Type type, u32 setting) const {
+u8 Mgr::GetSettingValue(Type type, u64 setting) const {
     return this->rawBin->GetSection<PagesHolder>().pages[type].settings[setting];
 }
-u8 Mgr::GetUserSettingValue(UserType type, u32 setting) const {
+u8 Mgr::GetUserSettingValue(UserType type, u64 setting) const {
     const PagesHolder& pagesHolder = this->rawBin->GetSection<PagesHolder>();
     return pagesHolder.pages[type + pagesHolder.pulsarPageCount].settings[setting];
 }
 
-void Mgr::SetSettingValue(Type type, u32 setting, u8 value) {
+void Mgr::SetSettingValue(Type type, u64 setting, u8 value) {
     this->rawBin->GetSection<PagesHolder>().pages[type].settings[setting] = value;
 }
-void Mgr::SetUserSettingValue(UserType type, u32 setting, u8 value) {
+void Mgr::SetUserSettingValue(UserType type, u64 setting, u8 value) {
     PagesHolder& pagesHolder = this->rawBin->GetSection<PagesHolder>();
     pagesHolder.pages[type + pagesHolder.pulsarPageCount].settings[setting] = value;
 }
